Feature/handle exceptions back

See merge request s09-fintech-finance-sub2/S09P22A207!59

+++ server/src/main/java/com/galaxy/memorit/friend/application/service/implement/FriendServiceImpl.java
@@ -9,11 +9,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.galaxy.memorit.common.exception.NoSuchFriendException;
+import com.galaxy.memorit.common.exception.NoSuchUserException;
 import com.galaxy.memorit.friend.Infrastructure.persistence.entity.FriendEntity;
 import com.galaxy.memorit.friend.Infrastructure.persistence.repository.FriendRepository;
 import com.galaxy.memorit.friend.Infrastructure.persistence.mapper.FriendMapper;
 import com.galaxy.memorit.friend.application.service.FriendService;
-import com.galaxy.memorit.friend.domain.entity.Friend;
 import com.galaxy.memorit.friend.dto.request.FriendMultiDeleteReqDTO;
 import com.galaxy.memorit.friend.dto.request.FriendRegisterFromAddressReqDTO;
 import com.galaxy.memorit.friend.dto.request.FriendRegisterReqDTO;
@@ -22,6 +23,7 @@ import com.galaxy.memorit.friend.dto.request.FriendUpdateReqDTO;
 import com.galaxy.memorit.friend.dto.response.FriendInfoDTO;
 import com.galaxy.memorit.friend.dto.response.FriendRankResDTO;
 import com.galaxy.memorit.friend.dto.response.FriendsListResDTO;
+import com.galaxy.memorit.user.domain.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -30,31 +32,37 @@ import lombok.RequiredArgsConstructor;
 public class FriendServiceImpl implements FriendService {
 	private final FriendRepository friendRepository;
 	private final FriendMapper friendMapper;
+	private final UserRepository userRepository;
 	@Transactional
 	@Override
 	public void registerFriend(String userId, FriendRegisterReqDTO dto) {
-		Friend friend = Friend.builder()
-			.userId(userId)
+		UUID userUUID = friendMapper.stringToUUID(userId);
+		//userId에 해당하는 회원이 존재하지 않을 때 예외 처리
+		userRepository.findById(userUUID).orElseThrow(NoSuchUserException::new);
+
+		FriendEntity friendEntity = FriendEntity.builder()
+			.friendId(UUID.randomUUID())
+			.userId(userUUID)
 			.name(dto.getName())
 			.category(dto.getCategory())
 			.build();
 
-		//String 형태인 userId를 UUID로 변환하여 저장
 		//friendId에 새로운 UUID 생성하여 저장
-		friendRepository.save(friendMapper.createEntity(friend));
+		friendRepository.save(friendEntity);
 	}
 
 	@Transactional
 	@Override
 	public void registerFriendsFromAddress(String userId, FriendRegisterFromAddressReqDTO dto) {
+		UUID userUUID = friendMapper.stringToUUID(userId);
+		userRepository.findById(userUUID).orElseThrow(NoSuchUserException::new);
+
 		List<FriendEntity> list = dto.getNameList().stream()
-				.map(name -> {
-					Friend friend = Friend.builder()
-						.userId(userId)
-						.name(name)
-						.build();
-					return friendMapper.createEntity(friend);
-				})
+				.map(name -> FriendEntity.builder()
+					.friendId(UUID.randomUUID())
+					.userId(userUUID)
+					.name(name)
+					.build())
 				.collect(Collectors.toList());
 		friendRepository.saveAll(list);
 	}
@@ -63,7 +71,10 @@ public class FriendServiceImpl implements FriendService {
 	@Override
 	public FriendsListResDTO getFriendsList(String userId) {
 		//byte[]를 UUID로 변경해서 db 조회
-		List<FriendEntity> entityList = friendRepository.findAllByUserId(friendMapper.stringToUUID(userId));
+		UUID userUUID = friendMapper.stringToUUID(userId);
+		userRepository.findById(userUUID).orElseThrow(NoSuchUserException::new);
+
+		List<FriendEntity> entityList = friendRepository.findAllByUserId(userUUID);
 
 		//db에서 얻은 리스트를 DTO에 맞게 변환
 		List<FriendInfoDTO> infoList = entityList.stream()
@@ -76,30 +87,51 @@ public class FriendServiceImpl implements FriendService {
 	@Transactional(readOnly = true)
 	@Override
 	public FriendInfoDTO getFriendInfo(String userId, String friendId) {
-		FriendEntity entity = friendRepository.findByFriendIdAndUserId(friendMapper.stringToUUID(friendId), friendMapper.stringToUUID(userId));
-		/*
-		entity 존재하는지 확인 필요. 없으면 not found
-		 */
+		UUID userUUID = friendMapper.stringToUUID(userId);
+		userRepository.findById(userUUID).orElseThrow(NoSuchUserException::new);
+
+		FriendEntity entity = friendRepository.findByFriendIdAndUserId(friendMapper.stringToUUID(friendId), userUUID);
+		if(entity == null){
+			throw new NoSuchFriendException();
+		}
+
 		return friendMapper.toInfoDTO(entity);
 	}
 
 	@Transactional
 	@Override
 	public void updateFriendInfo(String userId, String friendId, FriendUpdateReqDTO dto) {
-		FriendEntity entity = friendRepository.findByFriendIdAndUserId(friendMapper.stringToUUID(friendId), friendMapper.stringToUUID(userId));
+		UUID userUUID = friendMapper.stringToUUID(userId);
+		userRepository.findById(userUUID).orElseThrow(NoSuchUserException::new);
+
+		FriendEntity entity = friendRepository.findByFriendIdAndUserId(friendMapper.stringToUUID(friendId), userUUID);
+		if(entity == null){
+			throw new NoSuchFriendException();
+		}
+
 		entity.updateInfo(dto.getName(), dto.getCategory());
 	}
 
 	@Transactional
 	@Override
 	public void deleteFriendById(String userId, String friendId) {
-		FriendEntity entity = friendRepository.findByFriendIdAndUserId(friendMapper.stringToUUID(friendId), friendMapper.stringToUUID(userId));
+		UUID userUUID = friendMapper.stringToUUID(userId);
+		userRepository.findById(userUUID).orElseThrow(NoSuchUserException::new);
+
+		FriendEntity entity = friendRepository.findByFriendIdAndUserId(friendMapper.stringToUUID(friendId), userUUID);
+		if(entity == null){
+			throw new NoSuchFriendException();
+		}
+
 		friendRepository.delete(entity);
 	}
 
 	@Transactional
 	@Override
 	public void deleteFriendsByList(String userId, FriendMultiDeleteReqDTO dto) {
+		UUID userUUID = friendMapper.stringToUUID(userId);
+		userRepository.findById(userUUID).orElseThrow(NoSuchUserException::new);
+
 		//요청 받은 string list를 uuid list로 변환
 		List<UUID> uuidList = dto.getIdList().stream()
 				.map(friendMapper::stringToUUID)
@@ -109,8 +141,10 @@ public class FriendServiceImpl implements FriendService {
 
 	@Override
 	public FriendRankResDTO getFriendsRank(String userId) {
-		Pageable pageable = PageRequest.of(0,1);
 		UUID userUUID = friendMapper.stringToUUID(userId);
+		userRepository.findById(userUUID).orElseThrow(NoSuchUserException::new);
+
+		Pageable pageable = PageRequest.of(0,1);
 
 		List<FriendEntity> mostReceivedCountList = friendRepository.findFirstByReceivedCount(userUUID, pageable);
 		List<FriendEntity> mostSentCountList = friendRepository.findFirstBySentCount(userUUID, pageable);
@@ -133,7 +167,10 @@ public class FriendServiceImpl implements FriendService {
 
 	@Override
 	public FriendsListResDTO searchFriends(String userId, FriendSearchReqDTO dto) {
-		List<FriendEntity> entityList = friendRepository.findFriendsByDTO(friendMapper.stringToUUID(userId), dto);
+		UUID userUUID = friendMapper.stringToUUID(userId);
+		userRepository.findById(userUUID).orElseThrow(NoSuchUserException::new);
+
+		List<FriendEntity> entityList = friendRepository.findFriendsByDTO(userUUID, dto);
 
 		List<FriendInfoDTO> infoList = entityList.stream()
 			.map(friendMapper::toInfoDTO)
